+++ src/components/advanced-search/components/Search/components/SearchInput/components/index.ts
@@ -1,0 +1,8 @@
+export { TextInput } from './TextInput';
+export { EnumInput } from './EnumInput';
+export { NumberInput } from './NumberInput';
+export { DateInputGroup } from './DateInput';
+export {
+  InputSubmitButton,
+  type InputSubmitButtonProps,
+} from './InputSubmitButton';