+++ src/components/advanced-search/components/SortableWithCombine/components/SortableTreeItem.tsx
@@ -1,0 +1,60 @@
+import React, { CSSProperties } from 'react';
+import type { UniqueIdentifier } from '@dnd-kit/core';
+import { AnimateLayoutChanges, useSortable } from '@dnd-kit/sortable';
+import { CSS } from '@dnd-kit/utilities';
+import { TreeItem, TreeItemComponentProps } from './TreeItem';
+
+interface SortableTreeItemProps extends TreeItemComponentProps {
+  id: UniqueIdentifier;
+}
+
+const animateLayoutChanges: AnimateLayoutChanges = ({
+  isSorting,
+  wasDragging,
+}) => (isSorting || wasDragging ? false : true);
+
+/**
+ * [SortableTreeItem]:
+ * Handles the drag and drop functionality for the tree items.
+ */
+export function SortableTreeItem({
+  id,
+  depth,
+  ...props
+}: SortableTreeItemProps) {
+  const {
+    active,
+    attributes,
+    isDragging,
+    isSorting,
+    listeners,
+    setDraggableNodeRef,
+    setDroppableNodeRef,
+    transform,
+    transition,
+    ...rest
+  } = useSortable({
+    id,
+    animateLayoutChanges,
+  });
+  const style: CSSProperties = {
+    transform: CSS.Translate.toString(transform),
+    transition,
+  };
+  return (
+    <TreeItem
+      ref={setDraggableNodeRef}
+      wrapperRef={setDroppableNodeRef}
+      style={style}
+      depth={depth}
+      ghost={isDragging}
+      id={id}
+      disableInteraction={isSorting}
+      handleProps={{
+        ...attributes,
+        ...listeners,
+      }}
+      {...props}
+    />
+  );
+}