+++ src/components/resource-sections/components/dates/index.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
-import { Flex, FlexProps, Icon, Text } from 'nde-design-system';
+import { Flex, FlexProps, Icon, Link, Text } from 'nde-design-system';
 import { FaRegClock } from 'react-icons/fa';
 import TypeBanner from '../type-banner';
 import { FormattedResource } from 'src/utils/api/types';
+import { formatDOI } from 'src/utils/helpers';
 
 /*
 [COMPONENT INFO]:
@@ -36,6 +37,32 @@ export const DateTag: React.FC<DateTagProps> = ({ type, date, ...props }) => {
   );
 };
 
+export const DOITag = ({ doi, ...props }: { doi: string }) => {
+  return (
+    <Flex
+      px={2}
+      m={1}
+      alignItems='center'
+      bg='secondary.50'
+      w={['100%', '100%', 'unset']}
+      whiteSpace='nowrap'
+      borderRadius='semi'
+      {...props}
+    >
+      <Text fontSize='xs' mr={2}>
+        <strong>DOI</strong>
+      </Text>{' '}
+      {doi.includes('http') || doi.includes('doi.org') ? (
+        <Link href={doi} isExternal>
+          <Text fontSize='xs'>{formatDOI(doi)}</Text>
+        </Link>
+      ) : (
+        <Text fontSize='xs'>{formatDOI(doi)}</Text>
+      )}
+    </Flex>
+  );
+};
+
 interface ResourceDates {
   data?: FormattedResource;
 }
@@ -51,7 +78,6 @@ const ResourceDates: React.FC<ResourceDates> = ({ data }) => {
   if (!data) {
     return null;
   }
-
   const date_data = date_fields
     .map(({ property, ...fields }) => ({
       ...fields,
@@ -71,10 +97,12 @@ const ResourceDates: React.FC<ResourceDates> = ({ data }) => {
 
   return (
     <TypeBanner type={data?.type}>
-      <Flex flexWrap={'wrap'} ml={[0, 0, 4]}>
+      <Flex flexWrap='wrap' ml={[0, 0, 4]}>
         {date_data.map((date, i) => {
           return <DateTag key={i} type={date.name} date={date.value} />;
         })}
+
+        {data.doi && <DOITag doi={data.doi} />}
       </Flex>
     </TypeBanner>
   );