+++ src/components/search-results-page/components/sort/index.tsx
@@ -35,10 +35,10 @@ export const SortResults: React.FC<SortResultsProps> = ({
   */
   const getTooltip = (sortBy: SortOptionsInterface['sortBy']) => {
     if (sortBy === '_score') {
-      return 'Sort by relevancy.';
+      return 'Sort by relevancy (field name is boosted).';
     } else if (sortBy === 'date') {
       return 'Sort by most recent activity (created, published or modified).';
-    } else if (sortBy === 'name') {
+    } else if (sortBy === 'name.raw') {
       return 'Sort alphabetically by title.';
     }
     return '';