+++ src/pages/index.tsx
@@ -1,6 +1,7 @@
 import { useState } from 'react';
 import type { NextPage } from 'next';
 import {
+  Box,
   Button,
   Flex,
   Heading,
@@ -33,8 +34,8 @@ import {
 } from 'src/components/pie-chart';
 import { assetPrefix } from 'next.config';
 import NextLink from 'next/link';
-import { AdvancedSearch } from 'src/components/advanced-search';
 import { SearchBarWithDropdown } from 'src/components/search-bar';
+import { AdvancedSearchOpen } from 'src/components/advanced-search/components/buttons';
 
 const sample_queries = [
   {
@@ -114,16 +115,16 @@ const Home: NextPage = () => {
 
   interface Stats {
     datasets: Stat | null;
-    computationaltool: Stat | null;
+    // computationaltool: Stat | null;
     measurementTechnique: Stat | null;
     repositories: Stat | null;
   }
 
   const [stats, setStats] = useState<Stats>({
-    datasets: null,
-    computationaltool: null,
-    measurementTechnique: null,
     repositories: null,
+    datasets: null,
+    // computationaltool: null,
+    measurementTechnique: null,
   });
 
   const { isLoading, error } = useQuery<
@@ -169,16 +170,17 @@ const Home: NextPage = () => {
           stats: sources,
         };
         stat = {
-          datasets: types.dataset,
-          computationaltool: types.computationaltool,
-          measurementTechnique,
           repositories,
+          datasets: types.dataset,
+          // computationaltool: types.computationaltool,
+          measurementTechnique,
         };
       }
 
       setStats(stat);
     },
   });
+
   return (
     <>
       <PageContainer
@@ -194,25 +196,28 @@ const Home: NextPage = () => {
         >
           <>
             <Flex w='100%' justifyContent='flex-end' mb={2}>
-              <AdvancedSearch
-                buttonProps={{
-                  variant: 'outline',
-                  bg: 'whiteAlpha.500',
-                  color: 'white',
-                  _hover: { bg: 'whiteAlpha.800', color: 'primary.600' },
-                }}
-              />
+              <NextLink href={{ pathname: 'advanced-search' }} passHref>
+                <Box>
+                  <AdvancedSearchOpen
+                    onClick={() => {}}
+                    variant='outline'
+                    bg='whiteAlpha.500'
+                    color='white'
+                    _hover={{ bg: 'whiteAlpha.800', color: 'primary.600' }}
+                  />
+                </Box>
+              </NextLink>
             </Flex>
             <SearchBarWithDropdown
-              ariaLabel='Search for datasets or tools'
-              placeholder='Search for datasets or tools'
+              placeholder='Search for datasets'
+              ariaLabel='Search for datasets'
               size='md'
             />
 
             {/* [NOTE]: Test with autocomplete in the future */}
             {/* <SearchWithPredictiveText
-              ariaLabel='Search for datasets or tools'
-              placeholder='Search for datasets or tools'
+              ariaLabel='Search for datasets'
+              placeholder='Search for datasets'
               size='md'
               handleSubmit={(stringValue, __, data) => {
                 if (data && data.id) {
@@ -243,58 +248,18 @@ const Home: NextPage = () => {
                     }}
                     passHref
                   >
-                    <SearchQueryLink
-                      title={query.title}
-                      display={[i > 2 ? 'none' : 'block', 'block']}
-                    />
+                    <Box>
+                      <SearchQueryLink
+                        title={query.title}
+                        display={[i > 2 ? 'none' : 'block', 'block']}
+                      />
+                    </Box>
                   </NextLink>
                 );
               })}
             </Flex>
           </>
         </PageHeader>
-
-        {/* NIAID Data Ecosystem section */}
-        <PageContent justifyContent='center' bg='white' minH='unset'>
-          <StyledSection
-            id='nde'
-            alignItems='center'
-            flexDirection='column'
-            maxWidth={['100%', '100%', '750px']}
-          >
-            <StyledSectionHeading>
-              {homepageCopy.sections[1].heading}
-            </StyledSectionHeading>
-
-            <StyledText
-              textAlign={['start', 'center']}
-              mt={4}
-              fontSize={['lg', 'xl']}
-              lineHeight='taller'
-              maxW='unset'
-            >
-              {homepageCopy.sections[1].body}
-            </StyledText>
-            <StyledSectionButtonGroup variant='solid' justifyContent='center'>
-              {homepageCopy.sections[1]?.routes &&
-                homepageCopy.sections[1].routes.map(route => {
-                  return (
-                    <NextLink
-                      key={route.title}
-                      href={{
-                        pathname: route.path,
-                      }}
-                      passHref
-                    >
-                      <Button w='100%' my={2}>
-                        {route.title}
-                      </Button>
-                    </NextLink>
-                  );
-                })}
-            </StyledSectionButtonGroup>
-          </StyledSection>
-        </PageContent>
 
         {/* Display stats about the Biothings API */}
         {!error && (
@@ -398,11 +363,11 @@ const Home: NextPage = () => {
                 textAlign={['start', 'start', 'center', 'start']}
               >
                 <StyledSectionHeading mt={[4, 6]}>
-                  {homepageCopy.sections[2].heading}
+                  {homepageCopy.sections[1].heading}
                 </StyledSectionHeading>
-                <StyledText>{homepageCopy.sections[2].body}</StyledText>
-                {homepageCopy.sections[2]?.routes &&
-                  homepageCopy.sections[2].routes.map(
+                <StyledText>{homepageCopy.sections[1].body}</StyledText>
+                {homepageCopy.sections[1]?.routes &&
+                  homepageCopy.sections[1].routes.map(
                     (route: {
                       title: string;
                       path: string;