+++ src/utils/schema-definitions/fields.js
@@ -13,14 +13,33 @@ const fetchFields = async () => {
 
     const fields = Object.keys(fieldsData);
     let results = [];
+
+    // Add count for all fields option.
+    const allFields = await axios
+      .get('https://api-staging.data.niaid.nih.gov/v1/query/')
+      .then(response => {
+        return {
+          name: 'All Fields',
+          property: '',
+          count: response.data.total,
+          type: 'text',
+          description: '',
+        };
+      });
+    results.push(allFields);
+
     try {
-      results = await Promise.all(
+      const data = await Promise.all(
         fields.map(async property => {
           const response = await axios.get(
             `https://api-staging.data.niaid.nih.gov/v1/query?&q=_exists_:${property}`,
           );
-          const count = response.data.total;
+          let count = response.data.total;
 
+          // Skip @type property .
+          if (property === '@type') {
+            count = 0;
+          }
           return {
             name: getPropertyTitle(property),
             property,
@@ -29,6 +48,8 @@ const fetchFields = async () => {
           };
         }),
       );
+
+      results = [...results, ...data];
     } catch (e) {
       // do something to handle the error here
       console.log(e);
@@ -41,7 +62,7 @@ const fetchFields = async () => {
         // Filter data with needed types.
         return response.data.hits.filter(datum => {
           if (
-            datum.label === 'ComputationalTool' ||
+            // datum.label === 'ComputationalTool' ||
             datum.label === 'Dataset'
           ) {
             return datum.properties;
@@ -50,18 +71,29 @@ const fetchFields = async () => {
       });
 
     const data = [...ndeSchema].reduce((r, schemaData) => {
-      // Dataset or ComputationalTool
-      const resource_type = schemaData.label;
-      const fieldIndex = results.findIndex(f => f.property === '@type');
+      // // Dataset or ComputationalTool
+      const resource_type = 'Dataset';
 
-      // Add enum field for type property.
-      if (results[fieldIndex]) {
-        results[fieldIndex]['format'] = 'enum';
-        if (!results[fieldIndex]['enum']) {
-          results[fieldIndex]['enum'] = [];
-        }
-        results[fieldIndex]['enum'].push(resource_type);
-      }
+      // [NOTE] : This info is removed pending the return of tools to the API.
+      // const resource_type = schemaData.label;
+      // const fieldIndex = results.findIndex(f => f.property === '@type');
+
+      // // Add enum & description field for type property.
+      // if (results[fieldIndex]) {
+      //   results[fieldIndex]['format'] = 'enum';
+      //   if (!results[fieldIndex]['enum']) {
+      //     results[fieldIndex]['enum'] = [];
+      //   }
+      //   results[fieldIndex]['enum'].push(resource_type);
+      //   results[fieldIndex]['description'] = {
+      //     dataset: 'The type associated with the record.',
+      //     computationaltool: 'The type associated with the record (dataset).',
+      //   };
+      //   results[fieldIndex]['abstract'] = {
+      //     dataset: 'type of record (dataset)',
+      //     computationaltool: 'type of record (dataset)',
+      //   };
+      // }
 
       // Add property details to JSON
       const updateField = data => {