+++ CHANGELOG.md
@@ -1,3 +1,34 @@
 # Changelog
 
+All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
+
+### [0.0.2](https://github.com/NIAID-Data-Ecosystem/nde-portal/compare/v0.0.0...v0.0.2) (2023-02-27)
+
+
+### 📝 Documentation
+
+* update about page ([3e4e8f3](https://github.com/NIAID-Data-Ecosystem/nde-portal/commit/3e4e8f30b4d580aaff28622a16b8e2d64c17c16f))
+
+
+### 🐛 Bugfixes
+
+* [BREAKING CHANGE] results sorting with name.raw keyword ([f0d8df8](https://github.com/NIAID-Data-Ecosystem/nde-portal/commit/f0d8df80eb9b6d88b743ffa8bee5ea3d5e477208))
+* [BREAKING CHANGE] results sorting with name.raw keyword ([37a67d1](https://github.com/NIAID-Data-Ecosystem/nde-portal/commit/37a67d17f3ae9fddcbd421cd2c8c2fb9415271e1))
+* **filters:** limit date range + add date tags ([66cd095](https://github.com/NIAID-Data-Ecosystem/nde-portal/commit/66cd095a5e622ba11aaada7c40b74a31dc84081a))
+* **query-builder:** union bug ([57cac9e](https://github.com/NIAID-Data-Ecosystem/nde-portal/commit/57cac9ebb147af675c79aba7d4c94852e70a6a04))
+* sources date information ([bd10d06](https://github.com/NIAID-Data-Ecosystem/nde-portal/commit/bd10d06d70088945b85eacbd8195b0a3cfeb14f0))
+* sources date information ([230d2b1](https://github.com/NIAID-Data-Ecosystem/nde-portal/commit/230d2b15b5b5c7e784ea9800925fa67415ecdb71))
+
+
+### 🚀 Features
+
+* **pages:** add advanced search page ([fe3b7a1](https://github.com/NIAID-Data-Ecosystem/nde-portal/commit/fe3b7a108ee7e99dd40e276fa73dc6cfeac373ca))
+* **pages:** added faq + changelog pages ([b9efb4e](https://github.com/NIAID-Data-Ecosystem/nde-portal/commit/b9efb4e58caf56d659c0e003b2754c465752f379))
+* **pages:** added faq page ([518b8a5](https://github.com/NIAID-Data-Ecosystem/nde-portal/commit/518b8a57ae674fc5748071ffe93bfd8401eea91d))
+* **query-builder:** add search history ([2c0615a](https://github.com/NIAID-Data-Ecosystem/nde-portal/commit/2c0615a4366b382282f7d9924a90b3f96322f7f0))
+* **query-builder:** error handling ([b2ea480](https://github.com/NIAID-Data-Ecosystem/nde-portal/commit/b2ea4804a6897f6c209eaea126441f5f88c75f9f))
+* **query-builder:** update field config ([dac539a](https://github.com/NIAID-Data-Ecosystem/nde-portal/commit/dac539a5f5a1ccb49cec0d40eff16575d95bb694))
+* **search:** added search history ([8c08ba1](https://github.com/NIAID-Data-Ecosystem/nde-portal/commit/8c08ba1e8c94fcce5d857b939b4228d70b017ec8))
+* **search:** added search history toggle + added to dataset page ([cfd545c](https://github.com/NIAID-Data-Ecosystem/nde-portal/commit/cfd545c7fec7338cbbeaa553f3f50afba00fee2d))
+
 ### [0.0.1](https://github.com/NIAID-Data-Ecosystem/nde-portal/compare/v0.0.0...v0.0.1) (2023-01-31)